requirements.txt
Flask==2.2.2
python-dotenv==0.21.0
mysql-connector-python==8.0.32
pandas==1.5.3
Werkzeug==2.2.2

# Create a virtual environment (venv recommended)
python3 -m venv .venv

# Activate the virtual environment
# Linux/macOS:
source .venv/bin/activate
# Windows:
.venv\Scripts\activate

# Install from requirements.txt
pip install -r requirements.txt